audiointerleave: typecast bit-mask to guint64 to fix segmentation fault
While creating caps in audiointerleave tests, bitmask is being set as 0x9
This is resulting in segmentation fault. Fix the same by typecasting to guint64

https://bugzilla.gnome.org/show_bug.cgi?id=755840

+++ tests/check/elements/audiointerleave.c
@@ -1033,7 +1033,8 @@ GST_START_TEST (test_audiointerleave_2ch_smallbuf)
       "format", G_TYPE_STRING, GST_AUDIO_NE (F32),
       "channels", G_TYPE_INT, 2,
       "layout", G_TYPE_STRING, "interleaved",
-      "rate", G_TYPE_INT, 48000, "channel-mask", GST_TYPE_BITMASK, 0x9, NULL);
+      "rate", G_TYPE_INT, 48000, "channel-mask", GST_TYPE_BITMASK,
+      (guint64) 0x9, NULL);
 
   gst_event_parse_caps (ev, &ecaps);
   gst_check_caps_equal (ecaps, caps);
